==> chapters/2.CEA_Dimensionality_reduction_analysis.Rmd <==
---
title: "Cannabis Expression Atlas - Methods"
author: "Kevelin Barbosa-Xavier, Francisnei Pedrosa-Silva, Fabricio Almeida-Silva, Thiago M. Venancio"
date: "2024-08-22"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, eval = FALSE, warning = FALSE, message = FALSE)
```

## **Dimensionality reduction analysis**

**Resume:** Dimensionality reduction methods were used to evaluate sample clustering patterns. The ‘scran’ package (Lun, McCarthy and Marioni, 2016) was used to model the mean-variance relationship in a matrix of log-transformed counts normalized by library size, followed by the identification of the top 5000 genes with the highest expression variability. Next, we performed a principal component analysis (PCA) and used the elbow point statistic to extract the top principal components. Based on this, we employed the t-stochastic neighbor embedding (t-SNE) (Van Der Maaten and Hinton, 2008) dimensionality reduction method to compare the data distribution in a low-dimensional space. We tested 6 different perplexity values (10, 20, 30, 40, 50, and 60) and selected 40 as the optimal value based on visual inspection.

- Load necessary packages
```{r load_data}
set.seed(123) # for reproducibility
# Load packages
library(here)
library(bears)
library(GenomicFeatures)
library(SummarizedExperiment)
library(SingleCellExperiment)
library(scater)
library(scran)
library(DESeq2)
library(tidyverse)
library(patchwork)
library(ggplot2)
library(dplyr)
library(magrittr)
```

- choose the work directoy
```{r echo = F}
data <- "/media/winterfell/kevelin/doutorado/analises_JLmY_reference/CSEA_LATEST_08_2024/kevelin/"
```
```{r}
data <- data <- "/your/path/here/"
```

### **Step_1:** Feature selection

bears package note: "To maximize biological signal and reduce noise, we will only use highly
variable genes for dimensionality reduction. Here, we will pick the top 5000
of genes with the highest biological components." (Almeida-Silva, Pedrosa-Silva and Venancio, 2023).

```{r}
# load the data
load("data/se_atlas_gene.rda")

# Create a SingleCellExperiment with counts and log-normalized counts
atlas_counts_sce <- SingleCellExperiment(
    assays = list(
        counts = assay(se_atlas_gene, "gene_counts"),
        logcounts = log2(assay(se_atlas_gene, "gene_counts") + 1)
    ), 
    colData = colData(se_atlas_gene)
)
```

### **Step-2:** Modeling the mean-variance relationship and visualizing the fit
```{r}
mean_var_model <- modelGeneVar(atlas_counts_sce)
fit_mean_var <- metadata(mean_var_model)

plot_data <- data.frame(
    mean = fit_mean_var$mean,
    var = fit_mean_var$var,
    trend = fit_mean_var$trend(fit_mean_var$mean))
```

```{r}
#plot
p_fit_mean_var <- data.frame(
    plot_data) %>%
  ggplot(aes(x = mean, y = var)) +
  geom_point(alpha = 0.4) +
  geom_line(aes(y = trend), color = "steelblue3", linewidth = 1.5) +
  labs(
    title = "Per-gene mean-variance relationship",
    x = "Mean of log-expression", y = "Variance of log-expression"
    ) +
  theme_minimal() +
  theme(
        text = element_text(size = 14),
        axis.text = element_text(size = 14, color = "black"),
        axis.title = element_text(size = 14),
        plot.title = element_text(size = 16, face = "bold")
    )

p_fit_mean_var
```

```{r}
# Save plot
save(
    p_fit_mean_var, compress = "xz",
    file = here::here(data, "data/p_fit_mean_var.rda")
)
```

### **Step_3:** Extract the top 5000 of genes with the highest biological components
```{r}
hvg <- getTopHVGs(mean_var_model, n = 5000)
```

The object `hvg` is a character vector containing the IDs of the 
top 5000 genes with the highest biological components.

### **Step_4:** Principal components analysis (PCA)

Now, we will perform PCA using the genes in `hvg`.

- Identifying the Elbow Point
```{r}
# function for identify elbow point from explained variances
findElbowPoint <- function(variance) {
  if (is.unsorted(-variance)) {
    stop("'variance' should be sorted in decreasing order")
  }
  # Finding distance from each point on the curve to the diagonal.
  dy <- -diff(range(variance))
  dx <- length(variance) - 1
  l2 <- sqrt(dx^2 + dy^2)
  dx <- dx/l2
  dy <- dy/l2
  dy0 <- variance - variance[1]
  dx0 <- seq_along(variance) - 1
  parallel.l2 <- sqrt((dx0 * dx)^2 + (dy0 * dy)^2)
  normal.x <- dx0 - dx * parallel.l2
  normal.y <- dy0 - dy * parallel.l2
  normal.l2 <- sqrt(normal.x^2 + normal.y^2)
  # Picking the maximum normal that lies below the line.
  # If the entire curve is above the line, we just pick the last point.
  below.line <- normal.x < 0 & normal.y < 0
  if (!any(below.line)) {
      length(variance)
  } else {
      which(below.line)[which.max(normal.l2[below.line])]
  }
}
```

```{r}
# Perform PCA
atlas_counts_sce <- fixedPCA(
    atlas_counts_sce, subset.row = hvg
)

# Plot proportion of variance explained by each PC
percent_var <- attr(reducedDim(atlas_counts_sce), "percentVar")

sorted <- sort(percent_var, decreasing = T)
chosen_elbow <- findElbowPoint(sorted)
chosen_elbow
```

```{r}
# plot
p_pca_percent_var <- data.frame(
    Variance = round(percent_var, 2),
    PC = factor(1:50, levels = 1:20)
) |>
    ggplot(aes(x = PC, y = Variance)) +
    geom_col(fill = "grey40") +
    geom_text(aes(label = Variance), hjust = -0.3) +
    labs(
        title = "Proportion of variance explained by each PC",
        x = "PC", y = "Variance explained (%)"
    ) +
    coord_flip() +
    theme_minimal() +
    ylim(0, 25)+
  geom_vline(xintercept = chosen_elbow) +
  theme(
        text = element_text(size = 14),
        axis.text = element_text(size = 11, color = "black"),
        axis.title = element_text(size = 14),
        plot.title = element_text(size = 16, face = "bold")
    )

p_pca_percent_var

# Save plot
save(
    p_pca_percent_var, compress = "xz",
    file = here(data, "data/p_pca_percent_var.rda")
)
```

Based on the plot, we will use only the top 8 PCs for t-SNE and UMAP.

### **Step_5:** *t*-stochastic neighbor embedding (t-SNE)

Now, we will perform dimensionality reduction with t-SNE using the top 8 PCs
obtained previously. We will first test running a t-SNE with 6 different
perplexity values: 10, 20, 30, 40, 50, 60. Then, we will select the best.

```{r}
# Get and plot t-SNE coordinates (perplexity = 10, 20, 30, 40, 50)
perplexities <- c(10, 20, 30, 40, 50, 60)

p_tsne <- lapply(perplexities, function(x) {
    tsne_coord <- runTSNE(
        atlas_counts_sce, perplexity = x,
        dimred = "PCA", n_dimred = 8
    )
    # Color by the variable "Tissue"
    p <- plotReducedDim(tsne_coord, dimred = "TSNE", colour_by = "Tissue") +
        labs(
            x = "t-SNE 1", y = "t-SNE 2",
            title = paste0("Perplexity = ", x)
        )
    return(p)
})

# Visualize all plots
p_tsne_all_perplexities_panel <- wrap_plots(p_tsne, nrow = 2) +
    plot_layout(guides = "collect") &
    ggsci::scale_color_d3("category20") &
    labs(color = "Tissue")
```

Based on the plots, we chose `perplexity = 30` as the best option. Now, let's
create an object containing only the plot for this perplexity value.

- Plot t-SNE with perplexity = 30
```{r}
p_tsne_optimal_perplexity <- p_tsne_all_perplexities_panel[[3]] +
    labs(
        title = "t-SNE plot of all samples",
        subtitle = "Coordinates were constructed from the top 8 principal components, with perplexity = 30"
    )

# Save plots
save(
    p_tsne_optimal_perplexity, compress = "xz",
    file = here(data, "data/p_tsne_optimal_perplexity.rda")
)
```

```{r}
load("data/sample_metadata_new.rda")
sample_metadata_new$BioSample <- row.names(sample_metadata_new)
library(dplyr)
sample_metadata_new$Tissue[sample_metadata_new$Tissue == "Bast_fibre"] <- "Bast_Fibre"
# Create data frame
tsne_coordinates <- p_tsne_optimal_perplexity[["data"]]
tsne_coordinates <- as.data.frame(tsne_coordinates)
tsne_coordinates$BioSample <- rownames(tsne_coordinates)

tsne_coordinates  <- dplyr::rename(tsne_coordinates,
        tsne1 = X, 
        tsne2 = Y,
        Tissue = colour_by
    ) |>
    dplyr::select(tsne1, tsne2, BioSample, Tissue) |>
    dplyr::mutate(
        Tissue = str_to_title(Tissue),
        Tissue = as.factor(Tissue),
        BioSample = as.factor(BioSample)
    ) |>
    inner_join(
        sample_metadata_new |>
            dplyr::select(BioSample, Treatment, Cultivar, Chemotype)
    )

tsne_coordinates$Tissue <- gsub("Bast_fibre", "Bast_Fibre", tsne_coordinates$Tissue)

write.csv(tsne_coordinates,
          file = here(data, "tsne_coordinates.csv"))

# Save object
save(
    tsne_coordinates, compress = "xz",
    file = "data/tsne_coordinates.rda"
)
```
