==> chapters/7.CEA_paper_figures_tables.Rmd <==
---
title: "Cannabis Expression Atlas - Methods"
author: "Kevelin Barbosa-Xavier, Francisnei Pedrosa-Silva, Fabricio Almeida-Silva, Thiago M. Venancio"
date: "2024-08-22"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, eval = FALSE, warning = FALSE, message = FALSE)
```

```{r echo = FALSE, fig.align='center'}
data <- "/media/winterfell/kevelin/doutorado/analises_JLmY_reference/CSEA_LATEST_08_2024/kevelin/"
```
```{r}
# choose the work directoy
data <- "/your/path/here/"
```

## **CEA paper figures**

### Figure 1
- load data
```{r}
load("data/p_samples_per_Tissue.rda")
load("data/p_samples_per_chemotype.rda")
load("data/p_seqtech_layout_count.rda")
load("data/p_samples_per_Cultivar.rda")
```

```{r}
Figure_1_part_1 <- cowplot::plot_grid(
  p_samples_per_Tissue,
  p_samples_per_chemotype,
  p_seqtech_layout_count,
  nrow = 3,
  labels = c("(a)", "(b)", "(C)"),
  rel_heights = c(2, 1, 2)
)

save(Figure_1_part_1, compress = "xz",
     file = here(data, "data/Figure_1_part_1.rda"))
```

```{r}
Figure_1_part_2 <- p_samples_per_Cultivar
```

### Figure 2
- load data
```{r}
load("data/tsne_coordinates.rda")
load("data/annotation_colors_1.rda")
```

```{r}
    #tsne
   P_tsne_samples_tissue <- ggplot(
        tsne_coordinates, aes(
          x = tsne1, y = tsne2, sample = BioSample
        )
      ) + 
        geom_point(aes(color = Tissue), alpha = 0.9, size = 2) + 
     scale_color_manual(values = annotation_colors_1$Tissue) +
        theme_minimal() +
        labs(
          title = "t-sne plot of all samples colored by tissue",
          x = "t-SNE 1", y = "t-SNE 2"
        ) +
  theme(
        # Aumente o tamanho dos textos
    text = element_text(size = 14),
        axis.text = element_text(size = 11, color = "black"),
        axis.title = element_text(size = 14),
        plot.title = element_text(size = 16, face = "bold")
    ) +
     guides(fill = FALSE)


P_tsne_samples_type <- ggplot(
        tsne_coordinates, aes(
          x = tsne1, y = tsne2, sample = BioSample
        )
      ) + 
        geom_point(aes(colour = Chemotype), alpha = 0.9, size = 2) + 
        scale_color_manual(values = annotation_colors_1$Chemotype) +
        theme_minimal() +
        labs(
          title = "t-sne plot of all samples colored by chemotype",
          x = "t-SNE 1", y = "t-SNE 2"
        ) +
  theme(
        # Aumente o tamanho dos textos
    text = element_text(size = 14),
        axis.text = element_text(size = 11, color = "black"),
        axis.title = element_text(size = 14),
        plot.title = element_text(size = 16, face = "bold")
    )
```

```{r}
Figure_2 <- cowplot::plot_grid(
  P_tsne_samples_tissue, 
  P_tsne_samples_type,
  rel_widths = c(1.5, 1.35),
  nrow = 1, 
  labels = c("(a)", "(b)")
)

Figure_2

save(Figure_2, compress = "xz",
     file = here(data, "data/Figure_2.rda"))
```

### Figure 3
```{r}
load("data/p_genes_per_class.rda")
load("data/p_upset_genes_class.rda")
load("data/p_violin_tau.rda")
```

```{r}
Figure_3 <- cowplot::plot_grid(
  p_genes_per_class, 
  p_upset_genes_class,
  p_violin_tau,
  nrow = 3, 
  labels = c("(a)", "(b)", "(c)")
)

Figure_3
```

the upset plot can not be ploted with plot_grid, so we will plot it alone and using inkscape prepare the complete figure.
```{r}
Figure_3_upset <- p_upset_genes_class
```

### Figure_S1

```{r}
load("data/p_mapping_rate_density.rda")
load("data/p_q20_density.rda")
load("data/p_mean_length_density.rda")
load("data/p_nsamples_over_time_cumsum.rda")
load("data/p_Chemotype_timeseries.rda")
load("data/p_tissue_timeseries.rda")
load("data/p_barplot_samples_per_country.rda")
```

```{r}
p_samples_by_checkpoint <- data.frame(
  Initial = 515,
  Checkpoint_1 = 475,
  Checkpoint_2 = 394
)

p_samples_by_checkpoint <- t(p_samples_by_checkpoint)
p_samples_by_checkpoint <- as.data.frame(p_samples_by_checkpoint)
names(p_samples_by_checkpoint) <- "Number_of_samples"
p_samples_by_checkpoint$Steps <- rownames(p_samples_by_checkpoint)

# Set the "Number_of_samples" column as a factor with the desired order
p_samples_by_checkpoint$Steps <- factor(p_samples_by_checkpoint$Steps, levels = c("Checkpoint_2", "Checkpoint_1", "Initial" ))
# Sort dataframe based on "Steps" column in descending order
p_samples_by_checkpoint <- p_samples_by_checkpoint %>%
  arrange(desc(Number_of_samples))

# Plot data
library(ggplot2)
p_samples_by_checkpoint <- ggplot(
    p_samples_by_checkpoint, aes(x = Number_of_samples, y = Steps, fill = Number_of_samples)
) +
    geom_bar(stat = "identity") +
  geom_col(fill = "#053D38") +
    geom_text(aes(label = Number_of_samples), hjust = 1.2, size = 5) +  
    theme_minimal() +
    theme(
        
        text = element_text(size = 14),
        axis.text = element_text(size = 14, color = "black"),
        axis.title = element_text(size = 14),
        plot.title = element_text(size = 16, face = "bold")
    ) +
    labs(
      title = "Number of samples before and after filtering",
        x = "BioSamples", y = "",
        
    ) +
    scale_x_continuous(limits = c(0, 520)) +
    guides(fill = FALSE) 


p_samples_by_checkpoint
```

```{r}
Figure_S1 <- cowplot::plot_grid(
  p_samples_by_checkpoint, 
  p_mean_length_density,
  p_q20_density,
  p_mapping_rate_density,
  p_nsamples_over_time_cumsum,
  p_Chemotype_timeseries,
  p_tissue_timeseries,
  p_barplot_samples_per_country,
  nrow = 4, 
  labels = c("(a)", "(b)", "(c)", "(d)", "(e)", "(f)", "(g)", "(h)")
)

Figure_S1
```

### Figure_S2
```{r}
load("data/p_heatmap_HK.rda")
load("data/p_heatmap_median.rda")
```

```{r}
Figure_S2_part_1 <- p_heatmap_HK

Figure_S2_part_2 <- p_heatmap_median
```

## **CEA paper tables**

### Table 1 - TF per tissue
```{r}
load("data/TF_per_tissue.rda")

Table_1_TF_per_tissue <- TF_per_tissue
Table_1_TF_per_tissue$`TF family` <- row.names(Table_1_TF_per_tissue)

Table_1_TF_per_tissue[is.na(Table_1_TF_per_tissue)] <- ""

save(
    Table_1_TF_per_tissue,
    file = here::here(data, "data/0.tables/Table_1_TF_per_tissue.rda"),
    compress = "xz"
)

write_csv(Table_1_TF_per_tissue,
     file = here::here(data, "data/0.tables/Table_1_TF_per_tissue.csv"))
```

### Table 2 - Top 10 transcription factors family at Cannabis Expression Atlas compared to PlanTFDB.
Was constructed by hand, directly on the paper.

### Table S1 - sample metadata
```{r}
load("data/sample_metadata_new.rda")

Table_S1_sample_metadata <- sample_metadata_new

save(
    Table_S1_sample_metadata,
    file = here::here(data, "data/0.tables/Table_S1_sample_metadata.rda"),
    compress = "xz"
)

write_csv(Table_S1_sample_metadata,
     file = here::here(data, "data/0.tables/Table_S1_sample_metadata.csv"))
```

### Table S2 - Gene informations
```{r}
Table_S2_gene_informations <- CEA_annotation_latest

save(
    Table_S2_gene_informations,
    file = here::here(data, "data/0.tables/Table_S2_gene_informations.rda"),
    compress = "xz"
)

write_csv(Table_S2_gene_informations,
     file = here::here(data, "data/0.tables/Table_S2_gene_informations.csv"))
```

### Table S3 - n_genes_by_TF_family
```{r}
Table_S3_n_genes_by_TF_family <- n_genes_by_TF_family

save(
    Table_S3_n_genes_by_TF_family,
    file = here::here(data, "data/0.tables/Table_S3_n_genes_by_TF_family.rda"),
    compress = "xz"
)

write_csv(Table_S3_n_genes_by_TF_family,
     file = here::here(data, "data/0.tables/Table_S3_n_genes_by_TF_family.csv"))
```
